requirements.txt
pillow>=9.0.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0 
pydantic-ai